+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Validate port number when parsing 'http' URLs.
 - Replace "same authority" checks with "same origin".
 - Discard portable actor if ID of a special collection has different origin than actor ID.
+- Always use less verbose logging level for actix.
 
 ## [2.25.1] - 2024-07-11